+++ src/templates.simple/axosimple.app/Controllers/CultureController.cs
@@ -1,6 +1,7 @@
-﻿using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Localization;
+﻿using Microsoft.AspNetCore.Localization;
 using Microsoft.AspNetCore.Mvc;
+
+namespace axosimple.hmi.Controllers;
 
 [Route("/[controller]")]
 [ApiController]
