+++ src/templates.simple/axosimple.app/Controllers/ThemeController.cs
@@ -1,0 +1,15 @@
+﻿using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
+
+namespace axosimple.hmi.Controllers;
+
+[Route("/[controller]")]
+[ApiController]
+public class ThemeController : ControllerBase
+{
+    public async Task<ActionResult> ChangeTheme([FromQuery] string theme)
+    {
+        Response.Cookies.Append("theme", theme);
+        return Redirect("/");
+    }
+}
